orchestrator/requirements.txt
fastapi[all]
openai
tenacity
google-cloud-datastore
python-dotenv
